created components with separate pages/routes
