+++ app/components/footer/footer.js
