+++ app/components/introductions/howdie.js
@@ -1,0 +1,19 @@
+import Link from "next/link.js";
+
+export default function Howdie() {
+    return (
+        <>
+            {/* Introduction/Howdie */}
+            <section className="intro" id="home">
+                <h1 className="section__title section__title--intro">
+                    Hi, I am <strong>Jeremy Myers</strong>
+                </h1>
+                <p className="section__subtitle section__subtitle--intro">full-stack dev</p>
+                <img src="" alt="" />
+            </section>
+
+
+        </>
+
+    )
+}
