+++ app/page.js
@@ -1,13 +1,18 @@
 import Image from 'next/image'
-import Howdie from './howdie.js'
-import Dashboard from './dashboard/page.js'
-import Footer from './footer.js'
+import Howdie from './components/introductions/howdie.js'
+import Dashboard from './components/dashboard/page.js'
+import Footer from './components/footer/footer.js'
+import Services from './components/services/Services.js'
+import AboutMe from './components/about-me/AboutMe.js'
 
 export default function Home() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
       <div>
         <Howdie />
+        <Services />
+        <AboutMe />
+
       </div>
       <div className="z-10 w-full max-w-5xl items-center justify-between font-mono text-sm lg:flex">
         {/* <p className="fixed left-0 top-0 flex w-full justify-center border-b border-gray-300 bg-gradient-to-b from-zinc-200 pb-6 pt-8 backdrop-blur-2xl dark:border-neutral-800 dark:bg-zinc-800/30 dark:from-inherit lg:static lg:w-auto  lg:rounded-xl lg:border lg:bg-gray-200 lg:p-4 lg:dark:bg-zinc-800/30">
